load data for backlog item view
